feat(JAVA, JS): Made notification types show their icons and made drop down values show pretty versions of enums.
