+++ src/main/java/com/synopsys/integration/alert/provider/blackduck/descriptor/BlackDuckContent.java
@@ -118,7 +118,9 @@ public class BlackDuckContent extends ProviderContent {
         List.of(createStringField(createJsonPath(JsonField.FORMAT_DOUBLE_REPLACEMENT, JSON_FIELD_CONTENT, JSON_FIELD_BOM_COMPONENT), JSON_FIELD_BOM_COMPONENT, FieldContentIdentifier.TOPIC, LABEL_PROJECT_NAME),
             createStringField(createJsonPath(JsonField.FORMAT_DOUBLE_REPLACEMENT, JSON_FIELD_CONTENT, JSON_FIELD_BOM_COMPONENT), JSON_FIELD_BOM_COMPONENT, FieldContentIdentifier.SUB_TOPIC, LABEL_PROJECT_VERSION_NAME),
             createStringField(createJsonPath(JsonField.FORMAT_DOUBLE_REPLACEMENT, JSON_FIELD_CONTENT, JSON_FIELD_BOM_COMPONENT), JSON_FIELD_BOM_COMPONENT, FieldContentIdentifier.CATEGORY_ITEM, LABEL_BOM_COMPONENT)
-        ));
+        ),
+        "user-edit"
+        );
 
     public static final ProviderContentType LICENSE_LIMIT = new ProviderContentType(
         NotificationType.LICENSE_LIMIT.name(),
@@ -129,7 +131,8 @@ public class BlackDuckContent extends ProviderContent {
             createLongField(createJsonPath(JsonField.FORMAT_DOUBLE_REPLACEMENT, JSON_FIELD_CONTENT, JSON_FIELD_CODE_SIZE), JSON_FIELD_CODE_SIZE, FieldContentIdentifier.CATEGORY_ITEM, LABEL_LICENSE_LIMIT_USED_CODE_SIZE),
             createLongField(createJsonPath(JsonField.FORMAT_DOUBLE_REPLACEMENT, JSON_FIELD_CONTENT, JSON_FIELD_LIMIT_HARD), JSON_FIELD_LIMIT_HARD, FieldContentIdentifier.CATEGORY_ITEM, LABEL_LICENSE_LIMIT_HARD),
             createLongField(createJsonPath(JsonField.FORMAT_DOUBLE_REPLACEMENT, JSON_FIELD_CONTENT, JSON_FIELD_LIMIT_SOFT), JSON_FIELD_LIMIT_SOFT, FieldContentIdentifier.CATEGORY_ITEM, LABEL_LICENSE_LIMIT_SOFT)
-        )
+        ),
+        "gavel"
     );
 
     public static final ProviderContentType POLICY_OVERRIDE = new ProviderContentType(
@@ -156,7 +159,8 @@ public class BlackDuckContent extends ProviderContent {
             createObjectField(createJsonPath(JsonField.FORMAT_DOUBLE_REPLACEMENT, JSON_FIELD_CONTENT, JSON_FIELD_POLICY_INFOS), JSON_FIELD_POLICY_INFOS, FieldContentIdentifier.CATEGORY_ITEM, LABEL_POLICY_INFO_LIST,
                 new TypeRef<List<PolicyInfo>>() {})
 
-        )
+        ),
+        "exclamation-circle"
     );
 
     public static final ProviderContentType VULNERABILITY = new ProviderContentType(
@@ -186,7 +190,8 @@ public class BlackDuckContent extends ProviderContent {
             createObjectField(createJsonPath(JsonField.FORMAT_DOUBLE_REPLACEMENT, JSON_FIELD_CONTENT, JSON_FIELD_DELETED_VULNERABILITY_IDS), JSON_FIELD_DELETED_VULNERABILITY_IDS, FieldContentIdentifier.CATEGORY_ITEM,
                 LABEL_VULNERABILITY_DELETED,
                 new TypeRef<List<VulnerabilitySourceQualifiedId>>() {})
-        )
+        ),
+        "shield-alt"
     );
 
     private static final List<JsonField<?>> RULE_VIOLATION_FIELD_LIST = List.of(
@@ -205,11 +210,13 @@ public class BlackDuckContent extends ProviderContent {
     );
     public static final ProviderContentType RULE_VIOLATION = new ProviderContentType(
         NotificationType.RULE_VIOLATION.name(),
-        RULE_VIOLATION_FIELD_LIST
+        RULE_VIOLATION_FIELD_LIST,
+        "ban"
     );
     public static final ProviderContentType RULE_VIOLATION_CLEARED = new ProviderContentType(
         NotificationType.RULE_VIOLATION_CLEARED.name(),
-        RULE_VIOLATION_FIELD_LIST
+        RULE_VIOLATION_FIELD_LIST,
+        "eraser"
     );
 
     private static final Set<ProviderContentType> SUPPORTED_CONTENT_TYPES = Set.of(LICENSE_LIMIT, POLICY_OVERRIDE, RULE_VIOLATION, RULE_VIOLATION_CLEARED, VULNERABILITY, BOM_EDIT);
